charly | changes are being made so the manager will have a list of dao in order to manage the transactions
