+++ app/src/main/java/com/example/myapplication/database/dataManager/AccountManager.java
@@ -1,13 +1,21 @@
 package com.example.myapplication.database.dataManager;
 
+import android.database.Cursor;
+import android.database.sqlite.SQLiteDatabase;
+
+import com.example.myapplication.database.OpenHelper;
+import com.example.myapplication.database.dao.AccountDAO;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.example.myapplication.database.DataBaseContract.AccountEntry;
+
 public class AccountManager {
 
     private static AccountManager accountManager = null;
 
-    private Long id;
-    private String name;
-    private String description;
-    private Float balance;
+    private List<AccountDAO> accounts;
 
 
     private AccountManager() {
@@ -16,34 +24,30 @@ public class AccountManager {
     public static AccountManager getInstance() {
         if (accountManager == null) {
             accountManager = new AccountManager();
+            accountManager.accounts = new ArrayList<>();
         }
         return accountManager;
     }
 
 
-    public Long getId() {
-        return id;
-    }
-    public void setId(Long id) {
-        this.id = id;
-    }
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
-    public String getDescription() {
-        return description;
-    }
-    public void setDescription(String description) {
-        this.description = description;
-    }
-    public Float getBalance() {
-        return balance;
-    }
-    public void setBalance(Float balance) {
-        this.balance = balance;
+    public List<AccountDAO> getAll(OpenHelper dbHelper) {
+        SQLiteDatabase db = dbHelper.getReadableDatabase();
+        String[] columns = {AccountEntry._ID,
+                            AccountEntry.COLUMN_NAME,
+                            AccountEntry.COLUMN_DESCRIPTION,
+                            AccountEntry.COLUMN_BALANCE};
+
+        Cursor Accoutncur = db.query(AccountEntry.TABLE_NAME,
+                                     columns,
+                            null,
+                         null,
+                            null,
+                              null,
+                            AccountEntry.COLUMN_NAME + " DESC");
+
+        Accoutncur.close();/*TODO Delete this line and close the cursor properly when it is needed*/
+
+        return new ArrayList<AccountDAO>(); /*TODO put the proper return*/
     }
 
 }
